# backend/src/main/java/com/zl/mjga/dto/position/PositionQueryDto.java
package com.zl.mjga.dto.position;

import com.zl.mjga.model.urp.BindState;
import lombok.*;

@Getter
@Setter
@ToString
@AllArgsConstructor
@NoArgsConstructor
public class PositionQueryDto {
  private Long userId;
  private String name;
  private BindState bindState = BindState.ALL;
}