# frontend/src/api/schema/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/position": {
      "post": {
        "tags": [
          "position-controller"
        ],
        "operationId": "upsertPosition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "position-controller"
        ],
        "operationId": "deletePosition",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/user": {
      "get": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "queryUserWithRolePermission",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolePermissionDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "upsertUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpsertDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/role": {
      "get": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "queryRoleWithPermission",
        "parameters": [
          {
            "name": "roleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "upsertRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpsertDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "deleteRole",
        "parameters": [
          {
            "name": "roleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/role/unbind": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "unBindRoleBy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleBindDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/role/bind": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "bindRoleBy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleBindDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/position/unbind": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "unBindPositionBy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionBindDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/position/bind": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "bindPositionBy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionBindDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/permission": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "upsertPermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionUpsertDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "deletePermission",
        "parameters": [
          {
            "name": "permissionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/permission/unbind": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "unBindPermissionBy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionBindDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/permission/bind": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "bindPermissionBy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionBindDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/me": {
      "get": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "currentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolePermissionDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "upsertMe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpsertDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/department/unbind": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "unBindDepartmentBy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentBindDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/iam/department/bind": {
      "post": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "bindDepartmentBy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentBindDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/position/page-query": {
      "get": {
        "tags": [
          "position-controller"
        ],
        "operationId": "pageQueryPositions",
        "parameters": [
          {
            "name": "pageRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageRequestDto"
            }
          },
          {
            "name": "positionQueryDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PositionQueryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoListPositionRespDto"
                }
              }
            }
          }
        }
      }
    },
    "/iam/users": {
      "get": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "queryUsers",
        "parameters": [
          {
            "name": "pageRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageRequestDto"
            }
          },
          {
            "name": "userQueryDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserQueryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoListUserRolePermissionDto"
                }
              }
            }
          }
        }
      }
    },
    "/iam/roles": {
      "get": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "queryRoles",
        "parameters": [
          {
            "name": "pageRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageRequestDto"
            }
          },
          {
            "name": "roleQueryDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RoleQueryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoListRoleDto"
                }
              }
            }
          }
        }
      }
    },
    "/iam/permissions": {
      "get": {
        "tags": [
          "identity-access-controller"
        ],
        "operationId": "queryPermissions",
        "parameters": [
          {
            "name": "pageRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageRequestDto"
            }
          },
          {
            "name": "permissionQueryDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PermissionQueryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoListPermissionRespDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Position": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UserUpsertDto": {
        "required": [
          "enable",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "enable": {
            "type": "boolean"
          }
        }
      },
      "RoleUpsertDto": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "RoleBindDto": {
        "required": [
          "roleIds",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "PositionBindDto": {
        "required": [
          "positionIds",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "positionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "PermissionUpsertDto": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PermissionBindDto": {
        "required": [
          "permissionIds",
          "roleId"
        ],
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int64"
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "DepartmentBindDto": {
        "required": [
          "departmentIds",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "PageRequestDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "sortBy": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sortFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortFieldObject"
            }
          }
        }
      },
      "SortFieldObject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "order": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC",
              "DEFAULT"
            ]
          }
        }
      },
...
      "PositionQueryDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "bindState": {
            "type": "string",
            "enum": [
              "BIND",
              "UNBIND",
              "ALL"
            ]
          }
        }
      },
      "PageResponseDtoListPositionRespDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionRespDto"
            }
          }
        }
      },
      "PositionRespDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "isBound": {
            "type": "boolean"
          }
        }
      },
      "UserQueryDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
      "PageResponseDtoListUserRolePermissionDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRolePermissionDto"
            }
          }
        }
      },
      "PermissionRespDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isBound": {
            "type": "boolean"
          }
        }
      },
      "RoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isBound": {
            "type": "boolean"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionRespDto"
            }
          }
        }
      },
      "UserRolePermissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          },
          "enable": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "permissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionRespDto"
            }
          }
        }
      },
      "RoleQueryDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "roleId": {
            "type": "integer",
            "format": "int64"
          },
          "roleCode": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "roleIdList": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "bindState": {
            "type": "string",
            "enum": [
              "BIND",
              "UNBIND",
              "ALL"
            ]
          }
        }
      },
      "PageResponseDtoListRoleDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          }
        }
      },
      "PermissionQueryDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int64"
          },
          "permissionId": {
            "type": "integer",
            "format": "int64"
          },
          "permissionCode": {
            "type": "string"
          },
          "permissionName": {
            "type": "string"
          },
          "permissionIdList": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "bindState": {
            "type": "string",
            "enum": [
              "BIND",
              "UNBIND",
              "ALL"
            ]
          }
        }
      },
      "PageResponseDtoListPermissionRespDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionRespDto"
            }
          }
        }
      },
...
    }
  }
}